Improve navigation for business categories
First, add a NARROWER_THAN relationship to the graph to introduce
a hierarchy of business categories. Base the NARROWER_THAN relationship
on the Overlap Similarity algorithm in the Graph Data Science Library.
Then, modify the query that populates the Business Search drop down menu
to include only the top-level categories.