Improve the business search results
Order the business search results to include the most relevant
businesses for the "logged in" user. Use the Pearson similarity metric
and create a SIMILAR relationship between the user and other reviewers.
Then, order the businesses by the average score that users with the
SIMILAR relationship to the user gave to the business.